Extract msq logic from POSIX
