+++ qiling/os/posix/posix.py
@@ -3,7 +3,6 @@
 # Cross Platform and Multi Architecture Advanced Binary Emulation Framework
 #
 
-from collections import deque
 from inspect import signature, Parameter
 from typing import Dict, TextIO, Tuple, Union, Callable, IO, List, Optional
 
@@ -27,7 +26,7 @@ from qiling.cc import QlCC, intel, arm, mips, riscv, ppc
 from qiling.const import QL_ARCH, QL_OS, QL_INTERCEPT
 from qiling.exception import QlErrorSyscallNotFound
 from qiling.os.os import QlOs
-from qiling.os.posix.const import MSGMNB, NR_OPEN, errors
+from qiling.os.posix.msq import QlMsq
 from qiling.utils import ql_get_module, ql_get_module_function
 
 SYSCALL_PREF: str = f'ql_syscall_'
@@ -136,53 +135,6 @@ class QlShm:
     def get_by_attaddr(self, shmaddr: int) -> Optional[QlShmId]:
         return next((shmobj for shmobj in self.__shm.values() if shmobj.attach.count(shmaddr) > 0), None)
 
-
-class QlMsgBuf:
-    def __init__(self, mtype: int, mtext: bytes) -> None:
-        self.mtype = mtype
-        self.mtext = mtext
-
-
-# vaguely reflects a msqid64_ds structure
-class QlMsqId:
-    def __init__(self, key: int, uid: int, gid: int, mode: int) -> None:
-        # ipc64_perm
-        self.key = key
-        self.uid = uid
-        self.gid = gid
-        self.mode = mode
-
-        self.queue = deque(maxlen=MSGMNB)
-    
-    def __len__(self):
-        return len(self.queue)
-
-
-class QlMsq:
-    def __init__(self) -> None:
-        self.__msq: Dict[int, QlMsqId] = {}
-        self.__id: int = 0x0F000000
-
-    def __len__(self) -> int:
-        return len(self.__msq)
-
-    def add(self, msq: QlMsqId) -> int:
-        msqid = self.__id
-        self.__msq[msqid] = msq
-
-        self.__id += 0x1000
-
-        return msqid
-
-    def remove(self, msqid: int) -> None:
-        del self.__msq[msqid]
-
-    def get_by_key(self, key: int) -> Tuple[int, Optional[QlMsqId]]:
-        return next(((msqid, msqobj) for msqid, msqobj in self.__msq.items() if msqobj.key == key), (-1, None))
-
-    def get_by_id(self, msqid: int) -> Optional[QlMsqId]:
-        return self.__msq.get(msqid, None)
-    
 
 class QlOsPosix(QlOs):
